+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Features
 
 - Disable colorized output in case of `TERM=dumb` or `NO_COLOR=1`, see #542 and #555 (@nabijaczleweli)
+- Add new (experimental) `--min-benchmarking-time <secs>` option, see #527 (@sharkdp)
 
 ## Changes